requirements.txt
pytz==2022.1
requests==2.28.1
python-jose==3.3.0